feat(server): delete comment error handling

+++ packages/server/src/modules/comments/comments.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { CommentsRepository } from './comments.repository';
 
@@ -14,7 +14,11 @@ export class CommentsService {
     return await this.commentRepository.createComment(dto, authorId);
   }
 
-  async deleteComment(id: string) {
+  async deleteComment(id: string, userId: string) {
+    const comment = await this.commentRepository.findCommentById(id);
+    if (!comment) throw new HttpException('Comment not found', 404);
+    if (comment.userId !== userId)
+      throw new HttpException('You are not the author of this comment', 403);
     return await this.commentRepository.deleteComment(id);
   }
 }
